implementata mate_individuals ( ancora da perfezionare ) implementato costruttore di copia di individual completate alcuni commenti doxygen a bitmatrix.h e ga_engine.h

+++ clean.sh
@@ -1,12 +1,16 @@
 #!/bin/sh
 
-cp gentest gentest_tmp
+if [ -e gentest ] ; then
+    cp gentest gentest_tmp
+fi
 
 if [ -e "Makefile" ] ; then
     make clean
 fi
 
-mv gentest_tmp gentest
+if [ -e gentest_tmp ] ; then
+    mv gentest_tmp gentest
+fi
 
 find -name *~ -exec rm -f {} \;
 find -name *.o -exec rm -f {} \;
